requirements_rsna_privacy.txt
# Core ML stack
torch
torchvision

# Data handling & utils
numpy
pandas
tqdm
pillow

# DICOM support
pydicom
pylibjpeg
pylibjpeg-libjpeg

# Model evaluation
scikit-learn
